- Add a bool return value to FileSystem::ForEach function to stop ListFiles iteration when file/folder wanted has been found - Add "www.dir" server parameter - Redesign FileWatcher to watch folder/file changes with large featuring (sub folder, etc..) - Add a IOFile::watch feature to use FileWatcher in a asynchronous way - Redesign mona application server load: all application are loaded and main.lua changes update automatically the application and its children applications

+++ MonaBase/sources/FileLogger.cpp
@@ -62,11 +62,14 @@ void FileLogger::manage(UInt32 written) {
 	UInt32 maxNum(0);
 	Exception ex;
 	FileSystem::ForEach forEach([this, &ex, &name, &maxNum](const string& path, UInt16 level) {
-		UInt16 num = String::ToNumber<UInt16, 0>(FileSystem::GetBaseName(path, name));
+		UInt16 num;
+		if (!String::ToNumber(FileSystem::GetBaseName(path, name), num))
+			return true;
 		if (_rotation && num >= (_rotation - 1))
 			FileSystem::Delete(ex, String(_pFile->parent(), num, ".log"));
 		else if (num > maxNum)
 			maxNum = num;
+		return true;
 	});
 	FileSystem::ListFiles(ex, _pFile->parent(), forEach);
 	// rename log files
